backend/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.35
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.6
